+++ packages/launchdarkly-adapter/src/adapter.ts
