--- packages/launchdarkly-adapter/src/adapter/index.ts
@@ -1,1 +1,0 @@
-export { adapter } from './adapter';
